Same as r.9325 of FLTK 1.3: Fixed build error in msvc because ref to FULLSCREEN enum was not accessible in fltk3::Widget. New inline is_fullscreen() getter has been implemented to avoid a build error with (at least) msvc compilers. Fixed a ton of warnings / problems when building on windows 64 bits target with ms toolchain. cleaned up about 200 warnings raised when building win64 targets.
git-svn-id: file:///fltk/svn/branches/branch-3.0@9330 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fltk3/Browser.cxx
@@ -324,7 +324,7 @@ void fltk3::Browser::insert(int line, fltk3::BrowserLine_* item) {
   \param[in] d Optional pointer to user data to be associated with the new line.
 */
 void fltk3::Browser::insert(int line, const char* newtext, void* d) {
-  int l = strlen(newtext);
+  int l = (int)strlen(newtext);
   fltk3::BrowserLine_* t = (fltk3::BrowserLine_*)malloc(sizeof(fltk3::BrowserLine_)+l);
   t->length = (short)l;
   t->flags = 0;
@@ -359,7 +359,7 @@ void fltk3::Browser::move(int to, int from) {
 void fltk3::Browser::text(int line, const char* newtext) {
   if (line < 1 || line > lines) return;
   fltk3::BrowserLine_* t = find_line(line);
-  int l = strlen(newtext);
+  int l = (int)strlen(newtext);
   if (l > t->length) {
     fltk3::BrowserLine_* n = (fltk3::BrowserLine_*)malloc(sizeof(fltk3::BrowserLine_)+l);
     replacing(t, n);
