+++ src/fltk3/FileChooser2.cxx
@@ -906,7 +906,7 @@ fltk3::FileChooser::fileNameCB()
 
     // Other key pressed - do filename completion as possible...
     num_files  = fileList->size();
-    min_match  = strlen(filename);
+    min_match  = (int)strlen(filename);
     max_match  = min_match + 1;
     first_line = 0;
 
@@ -922,7 +922,7 @@ fltk3::FileChooser::fileNameCB()
 	if (!first_line) {
 	  // First match; copy stuff over...
 	  strlcpy(matchname, file, sizeof(matchname));
-	  max_match = strlen(matchname);
+	  max_match = (int)strlen(matchname);
 
           // Strip trailing /, if any...
 	  if (matchname[max_match - 1] == '/') {
@@ -960,14 +960,17 @@ fltk3::FileChooser::fileNameCB()
       fileList->redraw();
     } else if (max_match > min_match && first_line) {
       // Add the matching portion...
-      fileName->replace(filename - pathname, filename - pathname + min_match,
+      fileName->replace(
+			(int)(filename - pathname), 
+			(int)(filename - pathname + min_match),
                         matchname);
 
       // Highlight it with the cursor at the end of the selection so
       // s/he can press the right arrow to accept the selection
       // (Tab and End also do this for both cases.)
-      fileName->position(filename - pathname + max_match,
-	                 filename - pathname + min_match);
+      fileName->position(
+			 (int)(filename - pathname + max_match),
+	                 (int)(filename - pathname + min_match));
     } else if (max_match == 0) {
       fileList->deselect(0);
       fileList->redraw();
@@ -1356,7 +1359,7 @@ fltk3::FileChooser::update_preview()
 
     if (fp != NULL) {
       // Try reading the first 1k of data for a label...
-      bytes = fread(preview_text_, 1, sizeof(preview_text_) - 1, fp);
+      bytes = (int)fread(preview_text_, 1, sizeof(preview_text_) - 1, fp);
       preview_text_[bytes] = '\0';
       fclose(fp);
     } else {
@@ -1554,7 +1557,7 @@ fltk3::FileChooser::value(const char *filename)
   if (slash > pathname) slash[-1] = '/';
 
   fileName->value(pathname);
-  fileName->position(0, strlen(pathname));
+  fileName->position(0, (int)strlen(pathname));
   okButton->activate();
 
   // Then find the file in the file list and select it...
@@ -1621,8 +1624,8 @@ compare_dirnames(const char *a, const char *b) {
   int alen, blen;
 
   // Get length of each string...
-  alen = strlen(a) - 1;
-  blen = strlen(b) - 1;
+  alen = (int)strlen(a) - 1;
+  blen = (int)strlen(b) - 1;
 
   if (alen < 0 || blen < 0) return alen - blen;
