+++ src/fltk3/Menu_.cxx
@@ -37,7 +37,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#define SAFE_STRCAT(s) { len += strlen(s); if ( len >= namelen ) { *name='\0'; return(-2); } else strcat(name,(s)); }
+#define SAFE_STRCAT(s) { len += (int)strlen(s); if ( len >= namelen ) { *name='\0'; return(-2); } else strcat(name,(s)); }
 
 /** Get the menu 'pathname' for the specified menuitem.
 
@@ -89,7 +89,7 @@ int fltk3::Menu_::item_pathname(char *name, int namelen, const fltk3::MenuItem *
 		}
 	    } else {					// end of submenu? pop
 	        char *ss = strrchr(name, '/');
-		if ( ss ) { *ss = 0; len = strlen(name); }	// "File/Edit" -> "File"
+		if ( ss ) { *ss = 0; len = (int)strlen(name); }	// "File/Edit" -> "File"
 		else { name[0] = '\0'; len = 0; }		// "File" -> ""
 		continue;
 	    }
@@ -156,7 +156,7 @@ const fltk3::MenuItem * fltk3::Menu_::find_item(const char *pathname) {
 int fltk3::Menu_::find_index(const fltk3::MenuItem *item) const {
   fltk3::MenuItem *max = menu_+size();
   if (item<menu_ || item>=max) return(-1);
-  return(item-menu_);
+  return (int)(item-menu_);
 }
 
 /**
