+++ src/fltk3/encoding_mac_roman.cxx
@@ -87,7 +87,7 @@ static int n_buf = 0;
 
 const char *fltk3::local_to_mac_roman(const char *t, int n)  
 {
-  if (n==-1) n = strlen(t);
+  if (n==-1) n = (int)strlen(t);
   if (n<=n_buf) {
     n_buf = (n + 257) & 0x7fffff00;
     if (buf) free(buf);
@@ -108,7 +108,7 @@ const char *fltk3::local_to_mac_roman(const char *t, int n)
 
 const char *fltk3::mac_roman_to_local(const char *t, int n)
 {
-  if (n==-1) n = strlen(t);
+  if (n==-1) n = (int)strlen(t);
   if (n<=n_buf) {
     n_buf = (n + 257) & 0x7fffff00;
     if (buf) free(buf);
